added screen and object measurement components
