+++ src/types/MeasurementMode.ts
@@ -1,0 +1,1 @@
+export type MeasurementMode = 'screen' | 'object';
