+++ src/types/ObjectMeasurement.ts
@@ -1,0 +1,6 @@
+import { Vector3 } from 'three';
+
+export type ObjectMeasurement = {
+  normal: Vector3;
+  position: Vector3;
+};
